Use shared appendEscapedMarkdownInlineCode helper
Backslash-escaping backticks doesn't work inside a markdown inline
code span per  the only safe way is to choose a fence ofCommonMark
backticks longer than any run in the content (and pad with spaces
when the content begins/ends with a backtick).

The terminal tool already had this logic as a private
toMarkdownInlineCode helper. Promote it to htmlContent.ts as
appendEscapedMarkdownInlineCode (mirroring the existing
appendEscapedMarkdownCodeBlockFence helper for fenced blocks), use
it in copilotToolDisplay.ts, and switch sendToTerminalTool over to
the shared utility.

(Written by Copilot)
