+++ src/vs/platform/agentHost/node/copilot/copilotToolDisplay.ts
@@ -6,6 +6,7 @@
 import type { PermissionRequest } from '@github/copilot-sdk';
 import { hasKey } from '../../../../base/common/types.js';
 import { URI } from '../../../../base/common/uri.js';
+import { appendEscapedMarkdownInlineCode } from '../../../../base/common/htmlContent.js';
 import { localize } from '../../../../nls.js';
 import type { IAgentToolReadyEvent } from '../../common/agentService.js';
 import { StringOrMarkdown } from '../../common/state/protocol/state.js';
@@ -157,15 +158,6 @@ function truncate(text: string, maxLength: number): string {
 }
 
 /**
- * Escapes backticks in a value that will be embedded inside an inline-code
- * span (`` `...` ``) in a markdown string, so the value can't break out of
- * the code span.
- */
-function escapeMdInlineCode(value: string): string {
-	return value.replace(/`/g, '\\`');
-}
-
-/**
  * Formats a file path as a markdown link `[](file-uri)` so it renders
  * as a clickable file widget in the chat UI.
  */
@@ -212,7 +204,7 @@ export function getInvocationMessage(toolName: string, displayName: string, para
 		const args = parameters as ICopilotShellToolArgs | undefined;
 		if (args?.command) {
 			const firstLine = args.command.split('\n')[0];
-			return md(localize('toolInvoke.shellCmd', "Running {0}", '`' + escapeMdInlineCode(truncate(firstLine, 80)) + '`'));
+			return md(localize('toolInvoke.shellCmd', "Running {0}", appendEscapedMarkdownInlineCode(truncate(firstLine, 80))));
 		}
 		return localize('toolInvoke.shell', "Running {0} command", displayName);
 	}
@@ -242,14 +234,14 @@ export function getInvocationMessage(toolName: string, displayName: string, para
 		case CopilotToolName.Grep: {
 			const args = parameters as ICopilotGrepToolArgs | undefined;
 			if (args?.pattern) {
-				return md(localize('toolInvoke.grepPattern', "Searching for {0}", '`' + escapeMdInlineCode(truncate(args.pattern, 80)) + '`'));
+				return md(localize('toolInvoke.grepPattern', "Searching for {0}", appendEscapedMarkdownInlineCode(truncate(args.pattern, 80))));
 			}
 			return localize('toolInvoke.grep', "Searching files");
 		}
 		case CopilotToolName.Glob: {
 			const args = parameters as ICopilotGlobToolArgs | undefined;
 			if (args?.pattern) {
-				return md(localize('toolInvoke.globPattern', "Finding files matching {0}", '`' + escapeMdInlineCode(truncate(args.pattern, 80)) + '`'));
+				return md(localize('toolInvoke.globPattern', "Finding files matching {0}", appendEscapedMarkdownInlineCode(truncate(args.pattern, 80))));
 			}
 			return localize('toolInvoke.glob', "Finding files");
 		}
@@ -267,7 +259,7 @@ export function getPastTenseMessage(toolName: string, displayName: string, param
 		const args = parameters as ICopilotShellToolArgs | undefined;
 		if (args?.command) {
 			const firstLine = args.command.split('\n')[0];
-			return md(localize('toolComplete.shellCmd', "Ran {0}", '`' + escapeMdInlineCode(truncate(firstLine, 80)) + '`'));
+			return md(localize('toolComplete.shellCmd', "Ran {0}", appendEscapedMarkdownInlineCode(truncate(firstLine, 80))));
 		}
 		return localize('toolComplete.shell', "Ran {0} command", displayName);
 	}
@@ -297,14 +289,14 @@ export function getPastTenseMessage(toolName: string, displayName: string, param
 		case CopilotToolName.Grep: {
 			const args = parameters as ICopilotGrepToolArgs | undefined;
 			if (args?.pattern) {
-				return md(localize('toolComplete.grepPattern', "Searched for {0}", '`' + escapeMdInlineCode(truncate(args.pattern, 80)) + '`'));
+				return md(localize('toolComplete.grepPattern', "Searched for {0}", appendEscapedMarkdownInlineCode(truncate(args.pattern, 80))));
 			}
 			return localize('toolComplete.grep', "Searched files");
 		}
 		case CopilotToolName.Glob: {
 			const args = parameters as ICopilotGlobToolArgs | undefined;
 			if (args?.pattern) {
-				return md(localize('toolComplete.globPattern', "Found files matching {0}", '`' + escapeMdInlineCode(truncate(args.pattern, 80)) + '`'));
+				return md(localize('toolComplete.globPattern', "Found files matching {0}", appendEscapedMarkdownInlineCode(truncate(args.pattern, 80))));
 			}
 			return localize('toolComplete.glob', "Found files");
 		}
